restructure tileset processes and rename to reflect server destination
