+++ cloud_functions/data_processing/src/utils/mbtile_pipeline.py
@@ -17,7 +17,7 @@ logger = Logger()
 
 
 @dataclass
-class TilesetConfig:
+class MBTilesetConfig:
     bucket: str
     tileset_blob_name: str
     tileset_id: str
@@ -66,7 +66,7 @@ def _upload_mapbox(temp_dir: Path, ctx: dict[str, Any]):
 
 
 def run_tileset_pipeline(
-    cfg: TilesetConfig,
+    cfg: MBTilesetConfig,
     *,
     process: Hook | None = None,
     check_credentials: Callable = _check_mapbox_credentials,
